fix: Cleanup callbacks and move them to their own file
