+++ src/views/Dex/Migrate/MigrateAll/hooks/useStakeApproveAll.ts
@@ -4,7 +4,8 @@ import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { abi as IUniswapV2PairABI } from '@uniswap/v2-core/build/IUniswapV2Pair.json'
 import { useCallback } from 'react'
 import { getProviderOrSigner } from 'utils'
-import { ApeswapWalletLpInterface, MigrateStatus, useMigrateAll } from '../provider'
+import { ApeswapWalletLpInterface, MigrateStatus } from '../provider/types'
+import { useMigrateAll } from '../provider'
 import { useVaults } from 'state/vaults/hooks'
 import { useMasterChefAddress, useVaultApeAddressV2 } from 'hooks/useAddress'
