+++ src/views/Dex/Migrate/MigrateAll/provider/constants.ts
@@ -1,0 +1,7 @@
+export const MIGRATION_STEPS: { title: string; description: string }[] = [
+  { title: 'Unstake', description: 'some description' },
+  { title: 'Approve', description: 'some description' },
+  { title: 'Migrate', description: 'some description' },
+  { title: 'Approve', description: 'some description' },
+  { title: 'Stake', description: 'some description' },
+]
